requirements.txt
backcall==0.2.0
decorator==4.4.2
Django==2.2.13
djangorestframework==3.11.0
ipdb==0.13.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.4
